requirements-prod.txt
# Production requirements for ADocS service
# Core ML and NLP libraries (CPU-only versions)
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2

# API and HTTP libraries
anthropic==0.64.0
aiohttp==3.9.5
fastapi==0.110.3
uvicorn[standard]==0.29.0
pydantic==2.6.4

# Utilities
python-dotenv==1.0.1
huggingface-hub==0.24.6

# Additional dependencies for production
gunicorn==21.2.0
python-multipart==0.0.9

# Google Cloud Storage
google-cloud-storage==2.10.0
